update file handlers
